Fix using long

+++ vnext/ReactCommon/TEMP_UntilReactCommonUpdate/jsinspector-modern/NetworkIOAgent.h
@@ -36,8 +36,8 @@ struct LoadNetworkResourceRequest {
 
 struct ReadStreamParams {
   StreamID handle;
-  std::optional<unsigned long> size;
-  std::optional<unsigned long> offset;
+  std::optional<size_t> size;
+  std::optional<size_t> offset;
 };
 
 struct NetworkResource {
@@ -238,7 +238,7 @@ class NetworkIOAgent {
    * Stream IDs are strings of an incrementing integer, unique within each
    * NewtworkIOAgent instance. This stores the next one to use.
    */
-  unsigned long nextStreamId_{0};
+  size_t nextStreamId_{0};
 
   /**
    * Begin loading an HTTP resource, delegating platform-specific
